Move to XorShift64* RNG

+++ include/des.h
@@ -1,5 +1,7 @@
 #pragma once
+
 #include <stdbool.h>
+#include <stdint.h>
 
 struct result {
     double best_fit;
@@ -7,4 +9,4 @@ struct result {
     int count;
 };
 
-struct result des(int N, double function_fn(int N, double[N]), double lower[N], double upper[N], int seed, bool logRes);
+struct result des(int N, double function_fn(int N, double[N]), double lower[N], double upper[N], uint64_t seed, bool logRes);
